adding component to show the path and query parameters separately
